+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/TabsResponsiveToWindow.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import TabsTemplate from './TabsTemplate';
 
-export default () => (<TabsTemplate responsiveTo="window" />);
+export default () => (<TabsTemplate />);